+++ NEWS.md
@@ -1,6 +1,7 @@
 ## CHANGES IN audio.vadwebrtc VERSION 0.1
 
 - Added function VAD to detect voice in audio
+- Added is.voiced generic and is.voiced.webrtc-gmm to collapse voiced/non-voiced segments which are too short
 - Initial version based on webrtc commit 7f457533a2ee582865f50210e7460af90f78f0b6 (Wed Jan 17 19:03:20 2024)